Reordering inputs;
Changing exit.concern.spec.ts to not use pools JSON, and use TestPoolHelper instead;
Changing function that tests sorted arrays to exit.concern.spec.ts instead of integration;
